requirements.txt
Flask==3.1.0
gunicorn==22.0.0
# Change this:
# mysqlclient==2.2.0
# To this:
mysql-connector-python==8.4.0 # Or the latest stable version